requirements.txt
requests
bs4
markdown
newspaper3k
lxml_html_clean
python-json-logger
markitdown[all]